refactor: centralize canonical A2UI_UPDATE_KEYS in a2ui-bridge
Define canonical A2UI_UPDATE_KEYS in a2ui-bridge alongside
RenderA2uiItem as the single source of truth with JSDoc
documentation.

Reuse A2UI_UPDATE_KEYS in CrossFrameValidator and
surface-partitioner, documenting update key usage and isA2uiItem
streaming filtering.
